pyproject.toml
[project]
name = "task-manager-mcp"
version = "0.0.1"
description = "Task management MCP for Claude Code with dependency resolution, on top of Obsidian"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
